renamed integration_test_helper file (#240)
