+++ tests/nodes/supporting_nodes/test_file.py
@@ -4,7 +4,7 @@ import os
 import uuid
 
 import pytest
-from test_integration import integrate_nodes_helper
+from integration_test_helper import integrate_nodes_helper
 from util import strip_uid_from_dict
 
 import cript
